# openapi.json
{"openapi":"3.1.0","info":{"title":"Ticketing System","version":"0.1.0"},"paths":{"/payments/initiate":{"post":{"tags":["Payments"],"summary":"Initiate Payment","operationId":"initiate_payment_payments_initiate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentInitRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/payments/verify/{reference}":{"get":{"tags":["Payments"],"summary":"Verify Payment","operationId":"verify_payment_payments_verify__reference__get","parameters":[{"name":"reference","in":"path","required":true,"schema":{"type":"string","title":"Reference"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tickets/purchase":{"post":{"tags":["Tickets"],"summary":"Purchase Ticket","operationId":"purchase_ticket_tickets_purchase_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/tickets/mine":{"get":{"tags":["Tickets"],"summary":"My Tickets","operationId":"my_tickets_tickets_mine_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PaymentInitRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"amount":{"type":"integer","title":"Amount"}},"type":"object","required":["email","amount"],"title":"PaymentInitRequest"},"TicketRequest":{"properties":{"event_id":{"type":"integer","title":"Event Id"},"payment_reference":{"type":"string","title":"Payment Reference"}},"type":"object","required":["event_id","payment_reference"],"title":"TicketRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/users/login"}}}}}}